requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
einops==0.6.1
fastapi==0.95.0
filelock==3.10.7
h11==0.14.0
idna==3.4
llama-cpp-python==0.1.74
numpy==1.24.2
packaging==23.0
pydantic==1.10.7
PyYAML==6.0
requests==2.28.2
sentencepiece==0.1.97
sniffio==1.3.0
sse-starlette==1.3.3
starlette==0.26.1
tqdm==4.65.0
transformers==4.30.2
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
